Turned the old debug utils into logging utils
